requirements.txt
torch
torchvision
transformers
pandas
numpy
Pillow
scikit-learn
matplotlib
datasets
tqdm
